chore: move api env to deprecated
