+++ apps/api/_shared/schema.js
@@ -1,0 +1,4 @@
+// minimal schema validator placeholder
+exports.validate = function (_schema, _data) {
+  return true;
+};
