+++ apps/api/new/current.js
@@ -1,0 +1,3 @@
+module.exports = async function current(_req, res) {
+  res.json({ message: '현대인 스냅샷' });
+};
