+++ apps/web/cosmos/logo.png
