+++ deprecated/api_env
@@ -1,3 +1,4 @@
+# @deprecated
 # 서버 설정
 PORT=5000
 NODE_ENV=development
