+++ deprecated/chat.js
@@ -1,3 +1,4 @@
+// @deprecated
 const express = require('express');
 const router = express.Router();
