+++ deprecated/include.js
@@ -1,3 +1,4 @@
+// @deprecated
 // include.js : header.html include + 초기화
 window.addEventListener("DOMContentLoaded", async () => {
   const el = document.querySelector('[include-html]');
