+++ deprecated/news.js
@@ -1,3 +1,4 @@
+// @deprecated
 const express = require('express');
 const Parser = require('rss-parser');
 const router = express.Router();
